requirements.txt
streamlit
easyocr
pypdf
sentence-transformers
faiss-cpu
numpy
ollama
pymupdf
Pillow
torch